+++ CHANGELOG.md
@@ -10,15 +10,18 @@
 ### Added
 - Added an example for an addon of type `control`: addons\controls\exampleControl.
 - Added README.md for the Addon Cesium 3D Tiles Inspector.
+- Verkehrsstärken gfiTheme: Added Fullscreenview-button.
 
 ### Changed
 - packages: To avoid security issues, all packages have been fixed in package.json. No versions of patch releases will be accepted.
+- Verkehrsstärken gfiTheme: only shows data for the last 10 years.
 
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+- Verkehrsstärken gfiTheme: Fixed Anmerkung und Erhebungsmethode in tableview, fixed Baustelleneinfluss visibility in diagram, fixed layout.
 
 ---